+++ src/id_translation/fetching/_pandas_fetcher.py
@@ -1,4 +1,5 @@
-from collections.abc import Callable, Iterable, Mapping
+import logging
+from collections.abc import Callable, Mapping
 from pathlib import Path
 from typing import Any, Unpack
 
@@ -31,7 +32,8 @@ class PandasFetcher(AbstractFetcher[str, IdType]):
 
     Args:
         read_function: A Pandas `read`-function. String are resolved by :func:`~rics.misc.get_by_full_name`, using
-            ``default_module=pandas`` for convenience.
+            ``default_module=pandas`` for convenience. Will attempt to derive from `read_path_format` if ``None``
+            (requires a string path).
         read_path_format: A string on the form ``protocol://path/to/sources/{}.<ext>``, or a callable to apply
             to a source before passing them to `read_function`.
         read_function_kwargs: Additional keyword arguments for `read_function`.
@@ -45,13 +47,14 @@ class PandasFetcher(AbstractFetcher[str, IdType]):
 
     def __init__(
         self,
-        read_function: PandasReadFunction | str = "read_csv",
+        read_function: PandasReadFunction | str | None = None,
         read_path_format: str | FormatFn = "data/{}.csv",
         read_function_kwargs: Mapping[str, Any] | None = None,
         **kwargs: Unpack[AbstractFetcherParams[str, IdType]],
     ) -> None:
         super().__init__(**kwargs)
-        self._read = get_by_full_name(read_function, pd) if isinstance(read_function, str) else read_function
+
+        self._read = self._derive_read_function(read_function, read_path_format)
         self._format_source: FormatFn = read_path_format if callable(read_path_format) else read_path_format.format
         self._kwargs = read_function_kwargs or {}
 
@@ -94,12 +97,13 @@ class PandasFetcher(AbstractFetcher[str, IdType]):
         pattern = self.format_source("*")
 
         try:
-            source_paths = self._find_sources_fsspec(pattern)
+            sources = self._find_sources_fsspec(pattern)
         except ModuleNotFoundError as e:
             self.logger.debug(f"Falling back to 'pathlib.Path': {e!r}")
-            source_paths = self._find_sources_pathlib(pattern)
+            sources = self._find_sources_pathlib(Path(pattern).expanduser().resolve())
+        source_paths: dict[str, str] = {source: self.format_source(source) for source in sources}
 
-        extra = {"task_id": task_id, "pattern": pattern}
+        extra = {"task_id": task_id, "pattern": pattern, "source_paths": source_paths}
         if source_paths:
             self.logger.debug(f"Path {pattern=} matched {len(source_paths)} files: {source_paths}", extra=extra)
             return source_paths
@@ -107,31 +111,22 @@ class PandasFetcher(AbstractFetcher[str, IdType]):
             self.logger.warning(f"Path {pattern=} did not match any files.", extra=extra)
             return {}
 
-    def _find_sources_fsspec(self, pattern: str) -> dict[str, str]:
+    def _find_sources_fsspec(self, pattern: str) -> list[str]:
         from fsspec.core import url_to_fs  # type: ignore  # noqa: PLC0415
 
-        fs, _ = url_to_fs(pattern, **self._kwargs.get("storage_options", {}))
+        fs, pattern = url_to_fs(pattern, **self._kwargs.get("storage_options", {}))
+        return [self._path_to_source(path, pattern) for path in fs.glob(pattern)]
 
-        source_paths = self._make_source_paths(fs.glob(pattern))
+    @classmethod
+    def _find_sources_pathlib(cls, pattern: Path) -> list[str]:
+        return [cls._path_to_source(str(path), str(pattern)) for path in pattern.parent.glob(pattern.name)]
 
-        protocol, separator, _ = pattern.partition("://")
-        if separator:
-            source_paths = {source: protocol + separator + path for source, path in source_paths.items()}
-
-        return source_paths
-
-    def _find_sources_pathlib(self, pattern: str) -> dict[str, str]:
-        path = Path(pattern)
-        iterator = path.parent.glob(path.name)
-        return self._make_source_paths(iterator)
-
-    @staticmethod
-    def _make_source_paths(iterator: Iterable[Path]) -> dict[str, str]:
-        source_paths = {}
-        for path in map(Path, iterator):
-            source = path.with_suffix("").name
-            source_paths[source] = str(path)
-        return source_paths
+    @classmethod
+    def _path_to_source(cls, path: str, pattern: str) -> str:
+        glob_index = pattern.index("*")
+        prefix = pattern[:glob_index]
+        suffix = pattern[glob_index + 1 :]
+        return path.removeprefix(prefix).removesuffix(suffix)
 
     def _initialize_sources(self, task_id: int) -> dict[str, list[str]]:
         self._source_paths = self.find_sources(task_id)
@@ -146,3 +141,45 @@ class PandasFetcher(AbstractFetcher[str, IdType]):
     def __repr__(self) -> str:
         read_path_format = self.format_source("{}")
         return f"{tname(self)}(read_function={tname(self._read)}, {read_path_format=})"
+
+    def _derive_read_function(
+        self,
+        read_function: PandasReadFunction | str | None,
+        read_path_format: str | FormatFn,
+    ) -> PandasReadFunction:
+        if callable(read_function):
+            return read_function
+
+        if read_function is None:
+            if not isinstance(read_path_format, str):
+                msg = f"Cannot derive `read_function` from {read_path_format=} of type={type(read_function)}."
+                raise ValueError(msg)
+
+            func_to_suffixes: dict[PandasReadFunction, list[str]] = {
+                pd.read_csv: [".csv"],
+                pd.read_pickle: [".pickle", ".pkl"],
+                pd.read_feather: [".feather", ".ftr"],
+                pd.read_json: [".json"],
+                pd.read_parquet: [".parquet", ".parq"],
+            }
+
+            for func, suffixes in func_to_suffixes.items():
+                for suffix in suffixes:
+                    if suffix in read_path_format:  # Using endswith would break paths such as .csv.zip
+                        if self.logger.isEnabledFor(logging.INFO):
+                            read_function = f"pandas.{func.__name__}"
+                            self.logger.debug(
+                                f"Derived {read_function=} based on {suffix=} found in {read_path_format=}.",
+                                extra={"suffix": suffix, "read_function": read_function},
+                            )
+                        return func
+
+            suffixes = [suffix for suffixes in func_to_suffixes.values() for suffix in suffixes]
+            msg = f"Cannot derive `read_function` from {read_path_format=}; does not contain any known {suffixes=}."
+            raise ValueError(msg)
+
+        func = get_by_full_name(read_function, pd)
+        if not callable(func):
+            msg = f"Bad {read_function=}; type {type(func).__name__} is not callable."
+            raise TypeError(msg)
+        return func  # type: ignore[no-any-return]